Renamed use-action file
