+++ src/components/add-cell.tsx
@@ -1,0 +1,10 @@
+import './add-cell.css';
+import { useActions } from '../hooks/use-actions';
+
+interface AddCellProps {
+  nextCellId: string;
+}
+
+const AddCell: React.FC<AddCellProps> = ({ nextCellId }) => {
+  return <h1>Add Cell</h1>;
+};
